Move TypedDicts to separate file
