+++ internals/core_models.py
@@ -17,288 +17,13 @@
 # https://stackoverflow.com/a/33533514
 from __future__ import annotations
 
-from typing import Any, TypedDict
+from typing import Any
 
 from google.cloud import ndb  # type: ignore
 
 from framework import rediscache
 from internals.core_enums import *
 import settings
-
-
-class UserEditInfo(TypedDict):
-  by: str | None
-  when: str | None 
-
-class FeatureResourceInfo(TypedDict):
-  samples: list[str]
-  docs: list[str]
-
-
-class StandardsInfo(TypedDict):
-  spec: str | None
-  maturity: MaturityInfo
-
-
-class MaturityInfo(TypedDict):
-  text: str | None
-  short_text: str | None
-  val: int
-
-
-class BrowserStatus(TypedDict):
-  text: str | None
-  val: str | None
-  milestone_str: str | None
-
-
-class ViewInfo(TypedDict):
-  text: str | None
-  val: int | None
-  url: str | None
-  notes: str | None
-
-class ChromeBrowserInfo(TypedDict):
-  bug: str | None
-  blink_components: list[str] | None
-  devrel: list[str] | None
-  owners: list[str] | None
-  origintrial: bool | None
-  intervention: bool | None
-  prefixed: bool | None
-  flag: bool | None
-  status: BrowserStatus
-  desktop: int | None
-  android: int | None
-  webview: int | None
-  ios: int | None
-
-
-class SingleBrowserInfo(TypedDict):
-  view: ViewInfo | None
-
-
-class FeatureBrowsersInfo(TypedDict):
-  chrome: ChromeBrowserInfo
-  ff: SingleBrowserInfo
-  safari: SingleBrowserInfo
-  webdev: SingleBrowserInfo
-  other: SingleBrowserInfo
-
-
-class StageDict(TypedDict):
-  id: int
-  feature_id: int
-  stage_type: int
-  intent_stage: int
-  intent_thread_url: str | None
-
-  # Dev trial specific fields.
-  ready_for_trial_url: str | None
-  announcement_url: str | None
-
-  
-  # Origin trial specific fields.
-  experiment_goals: str | None
-  experiment_risks: str | None
-  extensions: list[StageDict]  # type: ignore
-  origin_trial_feedback_url: str | None
-  
-  # Trial extension specific fields.
-  ot_stage_id: int | None
-  experiment_extension_reason: str | None
-
-  # Ship specific fields
-  finch_url: str | None
-
-  # Enterprise specific fields
-  rollout_milestone: int | None
-  rollout_platforms: list[str]
-  rollout_details: list[str]
-  enterprise_policies: list[str]
-
-  # Email information
-  pm_emails: list[str]
-  tl_emails: list[str]
-  ux_emails: list[str]
-  te_emails: list[str]
-
-  # Milestone fields
-  desktop_first: int | None
-  android_first: int | None
-  ios_first: int | None
-  webview_first: int | None
-  desktop_last: int | None
-  android_last: int | None
-  ios_last: int | None
-  webview_last: int | None
-
-  # Legacy fields or fields that now live on stage entities.
-  # TODO(danielrsmith): stop representing these on the feature dict
-  # and references them direct from stage entities.
-  intent_to_implement_url: str | None
-  intent_to_experiment_url: str  | None
-  intent_to_extend_experiment_url: str | None
-  intent_to_ship_url: str | None
-
-  # Legacy milestone fields that now live on stage entities.
-  dt_milestone_desktop_start: int | None
-  dt_milestone_android_start: int | None
-  dt_milestone_ios_start: int | None
-  dt_milestone_webview_start: int | None
-  ot_milestone_desktop_start: int | None
-  ot_milestone_android_start: int | None
-  ot_milestone_webview_start: int | None
-  ot_milestone_desktop_end: int | None
-  ot_milestone_android_end: int | None
-  ot_milestone_webview_end: int | None
-  shipped_milestone: int | None
-  shipped_android_milestone: int | None
-  shipped_ios_milestone: int | None
-  shipped_webview_milestone: int | None
-
-
-class VerboseFeatureDict(TypedDict):
-
-  # Metadata: Creation and updates.
-  id: int
-  created: UserEditInfo
-  updated: UserEditInfo
-  accurate_as_of: str | None
-  creator_email: str | None
-  updater_email: str | None
-
-  # Metadata: Access controls
-  owner_emails: list[str]
-  editor_emails: list[str]
-  cc_emails: list[str]
-  spec_mentor_emails: list[str]
-  unlisted: bool
-  deleted: bool
-
-  # Renamed metadata fields
-  editors: list[str]
-  cc_recipients: list[str]
-  spec_mentors: list[str]
-  creator: str | None
-
-  # Descriptive info.
-  name: str
-  summary: str
-  category: str
-  category_int: int
-  blink_components: list[str]
-  star_count: int
-  search_tags: list[str]
-  feature_notes: str | None
-  enterprise_feature_categories: list[str]
-
-  # Metadata: Process information
-  feature_type: str
-  feature_type_int: int
-  intent_stage: str
-  intent_stage_int: int
-  active_stage_id: int | None
-  bug_url: str | None
-  launch_bug_url: str | None
-  breaking_change: bool
-
-  # Implementation in Chrome
-  flag_name: str | None
-  ongoing_constraints: str | None
-
-  # Topic: Adoption
-  motivation: str | None
-  devtrial_instructions: str | None
-  activation_risks: str | None
-  measurement: str | None
-  availability_expectation: str | None
-  adoption_expectation: str | None
-  adoption_plan: str | None
-
-  # Gate: Standardization and Interop
-  initial_public_proposal_url: str | None
-  explainer_links: list[str]
-  requires_embedder_support: bool
-  spec_link: str | None
-  api_spec: str | None
-  prefixed: bool | None
-  interop_compat_risks: str | None
-  all_platforms: bool | None
-  all_platforms_descr: bool | None
-  tag_review: str | None
-  non_oss_deps: str | None
-  anticipated_spec_changes: str | None
-
-  # Gate: Security & Privacy
-  security_risks: str | None
-  tags: list[str]
-  tag_review_status: str
-  tag_review_status_int: int | None
-  security_review_status: str
-  security_review_status_int: int | None
-  privacy_review_status: str
-  privacy_review_status_int: int | None
-
-  # Gate: Testing / Regressions
-  ergonomics_risks: str | None
-  wpt: bool | None
-  wpt_descr: str | None
-  webview_risks: str | None
-
-  # Gate: Devrel & Docs
-  devrel_emails: list[str]
-  debuggability: str | None
-  doc_links: list[str]
-  sample_links: list[str]
-
-  stages: list[StageDict]
-
-  # Legacy fields or fields that now live on stage entities.
-  # TODO(danielrsmith): stop representing these on the feature dict
-  # and references them direct from stage entities.
-  intent_to_implement_url: str | None
-  intent_to_experiment_url: str  | None
-  intent_to_extend_experiment_url: str | None
-  intent_to_ship_url: str | None
-  ready_for_trial_url: str | None
-  origin_trial_feeback_url: str | None
-  experiment_goals: str | None
-  experiment_risks: str | None
-  announcement_url: str | None
-  experiment_extension_reason: str | None
-  rollout_milestone: int | None
-  rollout_details: str | None
-  experiment_timeline: str | None
-  resources: FeatureResourceInfo
-  comments: str | None  # feature_notes
-
-  # Repeated in 'browsers' section. TODO(danielrsmith): delete these?
-  ff_views: int
-  safari_views: int
-  web_dev_views: int
-
-  browsers: FeatureBrowsersInfo
-
-  # Legacy milestone fields that now live on stage entities.
-  dt_milestone_desktop_start: int | None
-  dt_milestone_android_start: int | None
-  dt_milestone_ios_start: int | None
-  dt_milestone_webview_start: int | None
-  ot_milestone_desktop_start: int | None
-  ot_milestone_android_start: int | None
-  ot_milestone_webview_start: int | None
-  ot_milestone_desktop_end: int | None
-  ot_milestone_android_end: int | None
-  ot_milestone_webview_end: int | None
-
-  finch_url: str | None
-
-  standards: StandardsInfo
-  is_released: bool
-  is_enterprise_feature: bool
-  updated_display: str | None
-  new_crbug_url: str | None
 
 
 class FeatureEntry(ndb.Model):  # Copy from Feature
